// src/api/models/Creator.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jingle Jam Community Page Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Creator
 */
export interface Creator {
    /**
     * 
     * @type {string}
     * @memberof Creator
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Creator
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Creator
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof Creator
     */
    imageUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof Creator
     */
    color: string;
}

/**
 * Check if a given object implements the Creator interface.
 */
export function instanceOfCreator(value: object): value is Creator {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    if (!('color' in value) || value['color'] === undefined) return false;
    return true;
}

export function CreatorFromJSON(json: any): Creator {
    return CreatorFromJSONTyped(json, false);
}

export function CreatorFromJSONTyped(json: any, ignoreDiscriminator: boolean): Creator {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'url': json['url'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'color': json['color'],
    };
}

export function CreatorToJSON(value?: Creator | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'url': value['url'],
        'imageUrl': value['imageUrl'],
        'color': value['color'],
    };
}