// src/lib/analytics.ts
import {DateTime} from "luxon";
import type {Creator, Stream} from "../api";

export const log = (eventName: string, data: { [key: string]: any }) => {
  try {
    // @ts-ignore
    window.gtag('event', eventName, data)
  } catch (e) {
    console.error(e)
  }
}

export const logSlotClick = (slot: Stream) => {
  const start = DateTime.fromJSDate(slot.start)
  const data = {
    slot_title: slot.title,
    slot_year: start.year,
    event_label: `${start.year}_${start.day}_${start.hour}`,
  }
  log('click_slot', data)
}
export const logCreatorFromSlotClick = (creator: Creator, slot: Stream) => {
  const start = DateTime.fromJSDate(slot.start)
  const data = {
    slot_title: slot.title,
    slot_year: start.year,
    event_label: `${start.year}_${start.day}_${start.hour}`,
    name: creator.name,
  }
  log('click_creator_slot', data)
}
export const logCreatorSlotFilterClick = (creator: Creator, slot: Stream) => {
  const start = DateTime.fromJSDate(slot.start)
  const data = {
    slot_title: slot.title,
    slot_year: start.year,
    event_label: `${start.year}_${start.day}_${start.hour}`,
    name: creator.name,
  }
  log('creator_slot_filter', data)
}

export const logCreator = (creator: Creator) => {
  log('click_creator', {
    name: creator.name,
  })
}
